requirements.txt
streamlit
google-generativeai
PyMuPDF
Pillow
requests
pandas
fpdf
langchain-community
langchain-text-splitters
langchain-core
sentence-transformers